Added my own youtube directive, so I can specify the size and we see the controls again. Added slide style info to styles.css
IMPORTANT: I've changed the way of using tpl.html files!
Instead of converting them to a ts file, I simply use the templateUrl syntax in the directive (see slide.ts). In order for this to work, I had to remove the escape characters (needed for the conversion to ts) in the ng-class directive. Unfortunately, this means that all the templates would have to be retrieved in run time, which is potentially slow. Instead, I use a gulp-templatecache tool to populate the $templateCache in a run statement. The generated file is saved to includes/js/templates.js and is loaded, after the app.js so the app is initialized, in the index.html. Although not necessary during development, it does mean that the provided use-ref script also concatenates the templates into the techradar.js file in the debug-built folder.